+++ src/components/package-search-list.js
@@ -37,6 +37,7 @@ function PackageSearchList({
           onClick={() => onClickItem(`/eholdings/packages/${item.content.id}`)}
           showProviderName
           showTitleCount
+          showSelectedCount
         />
       )}
     />