+++ src/components/provider/show/provider-show.js
@@ -248,6 +248,8 @@ class ProviderShow extends Component {
         link={itemLink}
         item={item.content}
         showTitleCount
+        showSelectedCount
+        showTags
         headingLevel='h4'
       />
     );